Updated regex to get correct full capabilities URLs and flag so you can choose not to download new tarball each time.

+++ scripts/get_capability_urls_list.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 dir=`dirname $0`
 cd $dir/../data/
-rm data.gov.uk-ckan-meta-data-latest.csv.zip
-rm datasets.csv resources.csv
-wget https://data.gov.uk/data/dumps/data.gov.uk-ckan-meta-data-latest.csv.zip
-/usr/bin/unzip data.gov.uk-ckan-meta-data-latest.csv.zip
-egrep -Eio '(http|https)://[^/"].*?request=GetCapabilities' resources.csv | grep -v "," | sort | uniq > getCapabilitiesURLs.txt
+if [ "$1" == "download" ]; then
+	rm data.gov.uk-ckan-meta-data-latest.csv.zip
+	rm datasets.csv resources.csv
+	wget https://data.gov.uk/data/dumps/data.gov.uk-ckan-meta-data-latest.csv.zip
+	unzip data.gov.uk-ckan-meta-data-latest.csv.zip
+	rm getCapabilitiesURLs.txt
+fi
+egrep -Ei "(http|https)://" resources.csv | egrep -Ei "request=GetCapabilities" | egrep -Eio "(http|https)://[^,]*" | sort -u > getCapabilitiesURLs.txt
